@@ -5,6 +5,7 @@ import { gql } from "@apollo/client";
 import { print } from "graphql/language/printer";
 import type { Session } from "next-auth";
 import type { JWT } from "next-auth/jwt";
+import { env } from "../env.mjs";
 
 const LOGIN_USER = gql`
   mutation LoginUser($credentials: LoginInput!) {
@@ -73,7 +74,7 @@ export const authOptions: NextAuthOptions = {
         password: { label: "Password", type: "password" },
       },
       authorize: async (credentials) => {
-        const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/graphql`, {
+        const response = await fetch(`${env.API_URL}/graphql`, {
           method: "POST",
           headers: {
             "Content-Type": "application/json",
@@ -143,11 +144,11 @@ export const authOptions: NextAuthOptions = {
         httpOnly: true,
         sameSite: "lax",
         path: "/",
-        secure: process.env.NODE_ENV === "development",
+        secure: env.NODE_ENV === "development",
       },
     },
   },
-  secret: process.env.NEXTAUTH_SECRET,
+  secret: env.NEXTAUTH_SECRET,
 };
 
 async function refreshAccessToken(tokenObject: JWT) {
@@ -155,7 +156,7 @@ async function refreshAccessToken(tokenObject: JWT) {
   //console.log(tokenObject);
   try {
     // Get a new set of tokens with a refreshToken
-    const tokenResponse = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/graphql`, {
+    const tokenResponse = await fetch(`${env.API_URL}/graphql`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
@@ -189,5 +190,19 @@ async function refreshAccessToken(tokenObject: JWT) {
  * @see https://next-auth.js.org/configuration/nextjs
  */
 export const getServerAuthSession = (ctx: { req: GetServerSidePropsContext["req"]; res: GetServerSidePropsContext["res"] }) => {
-  return getServerSession(ctx.req, ctx.res, authOptions);
+  const session = getServerSession(ctx.req, ctx.res, authOptions);
+  if (!session) {
+    return {
+      redirect: {
+        destination: "/",
+        permanent: false,
+      },
+    };
+  }
+
+  return {
+    props: {
+      session,
+    },
+  };
 };
